#0: Passing rewrite of L-chip fast dispatch
