+++ tt_metal/hw/inc/dataflow_api.h
@@ -23,6 +23,9 @@
 #include "hostdevcommon/common_values.hpp"
 #include "risc_attribs.h"
 #include "third_party/umd/device/tt_silicon_driver_common.hpp"
+#ifndef COMPILE_FOR_ERISC
+#include "debug/dprint.h"
+#endif
 
 extern uint8_t noc_index;
 
@@ -146,6 +149,12 @@ void cb_push_back(const int32_t operand, const int32_t num_pages) {
     // producer always writes into contiguous memory, it cannot wrap
     if (cb_interface[operand].fifo_wr_ptr >= cb_interface[operand].fifo_limit) {
         // TODO: change this to fifo_wr_ptr
+        #ifndef COMPILE_FOR_ERISC
+        if (cb_interface[operand].fifo_wr_ptr > cb_interface[operand].fifo_limit) {
+            DPRINT << "BAD STATE" << ENDL();
+            while(true);
+        }
+        #endif
         cb_interface[operand].fifo_wr_ptr -= cb_interface[operand].fifo_size;
     }
 }
@@ -1609,6 +1618,8 @@ class Buffer {
         this->sharded = false;
     }
 
+    BufferType get_type() { return this->type; }
+
     void init_sharded(uint32_t page_size, uint32_t num_cores,  uint32_t addr, volatile tt_l1_ptr uint32_t* command_ptr){
         this->type = BufferType::L1;
         this->bank_base_address = addr;
@@ -1622,6 +1633,9 @@ class Buffer {
 
 
     void noc_async_write_buffer(uint32_t src, const uint32_t id, const uint32_t num_pages, const uint32_t offset=0) {
+        #ifndef COMPILE_FOR_ERISC
+        // DPRINT << "BUFFER WRITE" << ENDL();
+        #endif
         if (this->sharded) {
             noc_async_sharded_read_write_helper<false>(this->num_cores_, this->page_size_,
                                                 this->bank_base_address, this->base_command_addr_,
@@ -1632,8 +1646,20 @@ class Buffer {
                 noc_async_write(src, this->get_noc_addr_(id, offset), this->page_size_ * num_pages);
             }
             else {
+                // DPRINT << "BUF WRITE " << num_pages << " at " << (uint32_t)my_x[0] << ", " << (uint32_t)my_y[0] << ENDL();
+
+                // #ifndef COMPILE_FOR_ERISC
+                // #endif
                 for (uint32_t i = 0; i < num_pages; i++) {
                     uint64_t address = this->get_noc_addr_(id + i, offset);
+                    // #ifndef COMPILE_FOR_ERISC
+                    // DPRINT << "DRAM BUF WRITE" << ENDL();
+                    // uint32_t end = src + page_size_;
+                    // for (uint32_t i = src; i < end; i += sizeof(uint32_t)) {
+                    //     DPRINT << *reinterpret_cast<volatile tt_l1_ptr uint32_t*>(i) << ENDL();
+                    // }
+                    // DPRINT << ENDL();
+                    // #endif
                     noc_async_write(src, address, this->page_size_);
                     src += this->page_size_;
                 }
@@ -1643,6 +1669,9 @@ class Buffer {
     }
 
     void noc_async_read_buffer(uint32_t dst, const uint32_t id, const uint32_t num_pages, const uint32_t offset=0) {
+        #ifndef COMPILE_FOR_ERISC
+        // DPRINT << "BUFFER READ" << ENDL();
+        #endif
         if (this->sharded) {
             noc_async_sharded_read_write_helper<true>(this->num_cores_, this->page_size_,
                                             this->bank_base_address, this->base_command_addr_,
