+++ tt_metal/impl/dispatch/device_command.hpp
@@ -23,10 +23,8 @@ struct CommandHeader {
     uint32_t page_size = 0;
     uint32_t producer_cb_size = 0;
     uint32_t event;
-    uint32_t router_cb_size = 0;
     uint32_t consumer_cb_size = 0;
     uint32_t producer_cb_num_pages = 0;
-    uint32_t router_cb_num_pages = 0;
     uint32_t consumer_cb_num_pages = 0;
     uint32_t num_pages = 0;
     uint32_t num_runtime_arg_pages = 0;
@@ -38,8 +36,6 @@ struct CommandHeader {
     uint32_t issue_data_size = 0;
     uint32_t completion_data_size = 0;
     uint32_t producer_consumer_transfer_num_pages = 0;
-    uint32_t producer_router_transfer_num_pages = 0;
-    uint32_t consumer_router_transfer_num_pages = 0;
     uint32_t buffer_type = 0;
     uint32_t sharded_buffer_num_cores = 0;
     uint32_t restart = 0;
@@ -75,7 +71,6 @@ class DeviceCommand {
     static constexpr uint32_t NUM_ENTRIES_PER_BUFFER_TRANSFER_INSTRUCTION = COMMAND_PTR_SHARD_IDX + NUM_MAX_CORES*NUM_ENTRIES_PER_SHARD;
     static constexpr uint32_t NUM_POSSIBLE_BUFFER_TRANSFERS = 2;
     // Perf measurements showed best results with divisions of 4 pages being transferred from producer -> consumer
-    // TODO (abhullar): Sync with agrebenisan to replicate measurments for producer -> router -> consumer path
     static constexpr uint32_t SYNC_NUM_PAGES = 4;
 
     // Ensure any changes to this device command have asserts modified/extended
@@ -112,13 +107,9 @@ class DeviceCommand {
 
     void set_consumer_cb_size(const uint32_t cb_size);
 
-    void set_router_cb_size(const uint32_t cb_size);
-
     void set_producer_cb_num_pages(const uint32_t cb_num_pages);
 
     void set_consumer_cb_num_pages(const uint32_t cb_num_pages);
-
-    void set_router_cb_num_pages(const uint32_t cb_num_pages);
 
     void set_num_pages(const uint32_t num_pages);
 
@@ -139,10 +130,6 @@ class DeviceCommand {
     void set_completion_data_size(const uint32_t data_size);
 
     void set_producer_consumer_transfer_num_pages(const uint32_t producer_consumer_transfer_num_pages);
-
-    void set_producer_router_transfer_num_pages(const uint32_t producer_router_transfer_num_pages);
-
-    void set_consumer_router_transfer_num_pages(const uint32_t consumer_router_transfer_num_pages);
 
     uint32_t get_issue_data_size() const;
